move getMintApprovalSignature to utils
